modified 2004

+++ undefined_undefined/boj/math1/2004.cpp
@@ -2,9 +2,9 @@
 #include <algorithm>
 using namespace std;
 
-int countNum(int limit, const int target){
-	int cnt = 0;
-	for(int i = target; i <= limit; i *= target) cnt += (limit / i);
+long long countNum(int limit, const int target){
+	long long cnt = 0;
+	for(long long i = target; i <= limit; i *= target) cnt += (limit / i);
 	return cnt;
 }
 
@@ -15,7 +15,7 @@ int main(void){
 	 const int TWO = 2;
 	 int n, m;
 	 cin >> n >> m;
-	 int five_num = countNum(n, FIVE) - countNum(m, FIVE) - countNum(n - m, FIVE);
-	 int two_num = countNum(n, TWO) - countNum(m, TWO) - countNum(n - m, TWO);
+	 long long five_num = countNum(n, FIVE) - countNum(m, FIVE) - countNum(n - m, FIVE);
+	 long long two_num = countNum(n, TWO) - countNum(m, TWO) - countNum(n - m, TWO);
 	 cout << min(five_num, two_num);
 }
